final_project/app_prod/requirements.txt
attrs==24.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
mysqlclient==2.0.3
packaging==24.2
PyMySQL==1.1.1
python-dotenv==0.19.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
